+++ lib/tinkex/types/checkpoint_archive_url_response.ex
@@ -4,10 +4,11 @@ defmodule Tinkex.Types.CheckpointArchiveUrlResponse do
   """
 
   @type t :: %__MODULE__{
-          url: String.t()
+          url: String.t(),
+          expires: DateTime.t() | String.t() | nil
         }
 
-  defstruct [:url]
+  defstruct [:url, :expires]
 
   @doc """
   Convert a map (from JSON) to a CheckpointArchiveUrlResponse struct.
@@ -15,7 +16,20 @@ defmodule Tinkex.Types.CheckpointArchiveUrlResponse do
   @spec from_map(map()) :: t()
   def from_map(map) do
     %__MODULE__{
-      url: map["url"] || map[:url]
+      url: map["url"] || map[:url],
+      expires: parse_expires(map["expires"] || map[:expires])
     }
   end
+
+  defp parse_expires(nil), do: nil
+  defp parse_expires(%DateTime{} = dt), do: dt
+
+  defp parse_expires(value) when is_binary(value) do
+    case DateTime.from_iso8601(value) do
+      {:ok, dt, _offset} -> dt
+      _ -> value
+    end
+  end
+
+  defp parse_expires(other), do: other
 end